#52 Refactor for better reorganization and clarity
