+++ src/app/api/openai/threads/[id]/messages/route.ts
@@ -1,7 +1,7 @@
 import OpenAI from 'openai';
 import { PrismaClient } from '@prisma/client';
 import { NextRequest, NextResponse } from 'next/server';
-import { getOpenAIObjectForAssistant } from '@/app/api/utils';
+import { getOpenAIObjectForAssistant } from '@/app/api/openai/util';
 
 const prisma = new PrismaClient();
