+++ src/app/api/openai/threads/[id]/runs/route.ts
@@ -4,33 +4,9 @@ import { NextRequest, NextResponse } from 'next/server';
 import { Stream } from 'openai/streaming';
 import { ulid } from 'ulidx';
 import { createMessage } from '@/app/api/utils/messages';
+import { getOpenAIObjectForAssistant } from '@/app/api/openai/util';
 
 const prisma = new PrismaClient();
-
-export const getOpenAIObjectForAssistant = async (
-  req: NextRequest,
-  prisma: PrismaClient
-) => {
-  let assistantId = req.headers.get('X-Assistant-Id');
-
-  // @ts-ignore
-  let assistant = await prisma.assistant.findFirst({
-    where: {
-      id: assistantId ? assistantId : undefined,
-    },
-    include: {
-      organization: true,
-    },
-  });
-
-  if (!assistant) {
-    throw new Error('Assistant does not exist');
-  }
-
-  return new OpenAI({
-    apiKey: assistant?.organization?.openAIApiKey,
-  });
-};
 
 const getId = (req: Request) => {
   const url = new URL(req.url);
